No more INT

+++ output.h
@@ -7,12 +7,12 @@
 
 // Print results to console
 void printResults(const std::string &jobName, const std::string &tubing,
-                  int stockLen, double kerf, const std::vector<Cut> &cuts,
+                  double stockLen, double kerf, const std::vector<Cut> &cuts,
                   const Solution &solution);
 
 // Generate HTML output file
 void generateHTML(const std::string &filename, const std::string &jobName,
-                  const std::string &tubing, int stockLen, double kerf,
+                  const std::string &tubing, double stockLen, double kerf,
                   const std::vector<Cut> &cuts, const Solution &solution);
 
 // Group sticks into patterns for cleaner output
