+++ parse.h
@@ -4,13 +4,13 @@
 #include <string>
 
 // Parse various length formats (e.g., "24'", "8'4\"", "180 1/2", "288")
-int parseAdvancedLength(const std::string &s);
+double parseAdvancedLength(const std::string &s);
 
 // Parse a fraction or decimal (e.g., "1/2", "0.125", "3/16")
 double parseFraction(const std::string &s);
 
-// Format inches as feet and inches (e.g., 100 -> "8'4\"")
-std::string prettyLen(int inches);
+// Format inches as feet and inches (e.g., 100.5 -> "8' 4 1/2\"")
+std::string prettyLen(double inches);
 
 // Get user input with a default value
 std::string getInput(const std::string &prompt,
